cleaning up the HomeComponents
