+++ app/src/main/java/com/example/clicker/presentation/stream/views/chat/ImprovedChatUI.kt
@@ -135,7 +135,6 @@ import com.example.clicker.network.websockets.models.MessageToken
 
 import com.example.clicker.network.websockets.models.MessageType
 import com.example.clicker.network.websockets.models.PrivateMessageType
-import com.example.clicker.presentation.home.TestingGLSurfaceViewComposable
 import com.example.clicker.presentation.stream.util.FilteredChatterListImmutableCollection
 import com.example.clicker.presentation.stream.util.ForwardSlashCommandsImmutableCollection
